GUI change - Last optimizations

+++ src/de/thm/se/calculator/calc/Factorial.java
@@ -1,6 +1,7 @@
 package de.thm.se.calculator.calc;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 /**
  * A factorial is the product of all positive integers less than or equal to n
@@ -29,13 +30,13 @@ public class Factorial implements Calculatable {
 
 	@Override
 	public BigDecimal getValue() {
-		int n = this.n.getValue().intValue();
+		long n = this.n.getValue().longValue();
 		
 		if(n <= 1)
 			return BigDecimal.ONE;
 		
-        int r = 1;
-		for(int i = 2; i <= n; i++) r *= i;
+        long r = 1;
+		for(long i = 2; i <= n; i++) r *= i;
         return new BigDecimal(r);
 	}
